Exposes LockingCache configuration to adjust ttl and expire policy
LockingCache now accepts deleteOnHit option to delete objects when
they are requested even if they were set with a ttl/timeout > 0.

MapnikSource now accepts configuring ttl and timeout/ttl for the
internal locking cache so it's possible to combine it with
metatiling in a more efficient manner.

Now it's possible to use metatile, use a ttl>0 and remove on hit, so
cached objects from metatile will be removed after ttl but objects
already requested are removed as they get a hit.